Testing change to reduce over-discharge chance
Implementing test to remove the next hour, and current hour forecast from the forecast. This reduces the chance that the inverter will overdischarge (Discharge more than necessary, meaning the battery needs to be charged) during the evening. This change should have negligible impact on the morning charge forecast.

This is necessary because of the way that Solis inverters handle discharging whilst under "self-use" mode and schedules. During a "discharge" schedule, all solar generation and any battery power available - up to the maximum amperage set - will be sent directly to the grid. I believe the assumption made in this code originally is that the solar generation would continue to "top up" the battery power during this schedule - however, from experience this is not the case.